Extract FindAPIHelper for reuse
Make stub_200_from_find use kwargs so we can define (a bit) what it is
we're going to get back when we declare the stub
