+++ spec/support/test_helpers/find_api_helper.rb
@@ -1,0 +1,66 @@
+module FindAPIHelper
+  def stub_200_from_find(provider_code:, provider_name:)
+    stub_find_api_provider(provider_code)
+      .to_return(
+        status: 200,
+        headers: { 'Content-Type': 'application/vnd.api+json' },
+        body: {
+          'data': {
+            'id': '1',
+            'type': 'providers',
+            'attributes': {
+              'provider_name': provider_name,
+            },
+            'relationships': {
+              'sites': {
+                'data': [
+                  { 'id': '1', 'type': 'sites' },
+                ],
+              },
+              'courses': {
+                'data': [
+                  { 'id': '1', 'type': 'courses' },
+                ],
+              },
+            },
+          },
+          'included': [
+            {
+              'id': '1',
+              'type': 'sites',
+              'attributes': {
+                'location_name': '-',
+                'name': 'Main Site',
+              },
+            },
+            {
+              'id': '1',
+              'type': 'courses',
+              'attributes': {
+                'course_code': 'X130',
+                'name': 'Primary',
+                'level': 'primary',
+                'start_date': 'September 2019',
+              },
+              'relationships': {
+                'sites': {
+                  'data': [
+                    { 'id': '1', 'type': 'sites' },
+                  ],
+                },
+              },
+            },
+          ],
+          'jsonapi': { 'version': '1.0' },
+        }.to_json,
+      )
+  end
+
+private
+
+  def stub_find_api_provider(provider_code)
+    stub_request(:get, ENV.fetch('FIND_BASE_URL') +
+      'recruitment_cycles/2020' \
+      "/providers/#{provider_code}?include=sites,courses.sites")
+  end
+end
